fix functions

+++ potapova.marina/I4/countColsWithNull.h
@@ -3,6 +3,6 @@
 
 #include <cstddef>
 
-size_t countColsWithNull(const int* matrix, size_t rows, size_t cols);
+size_t countColsWithNull(const long long* matrix, size_t rows, size_t cols);
 
 #endif
