+++ potapova.marina/I4/countRowsWithoutNull.h
@@ -3,6 +3,6 @@
 
 #include <cstddef>
 
-size_t countRowsWithoutNull(const int* matrix, size_t count_rows, size_t count_cols);
+size_t countRowsWithoutNull(const long long* matrix, size_t count_rows, size_t count_cols);
 
 #endif
